refactor FilesTableFilters
