+++ src/sections/dataset/dataset-files/DatasetFiles.tsx
@@ -4,7 +4,7 @@ import { useFiles } from './useFiles'
 import { useEffect, useState } from 'react'
 import { FilesTable } from './files-table/FilesTable'
 import { SpinnerSymbol } from './files-table/spinner-symbol/SpinnerSymbol'
-import { FileCriteriaInputs } from './file-criteria-inputs/FileCriteriaInputs'
+import { FileCriteriaControls } from './file-criteria-controls/FileCriteriaControls'
 import { FileCriteria } from '../../../files/domain/models/FileCriteria'
 
 interface DatasetFilesProps {
@@ -44,7 +44,7 @@ export function DatasetFiles({
   return (
     <>
       {files.length >= MINIMUM_FILES_TO_SHOW_CRITERIA_INPUTS && (
-        <FileCriteriaInputs
+        <FileCriteriaControls
           criteria={criteria}
           onCriteriaChange={handleCriteriaChange}
           filesCountInfo={filesCountInfo}
