+++ src/sections/dataset/dataset-files/file-criteria-controls/FileCriteriaSortBy.tsx
@@ -1,5 +1,5 @@
 import { ArrowDownUp } from 'react-bootstrap-icons'
-import styles from './FileCriteriaInputs.module.scss'
+import styles from './FileCriteriaControls.module.scss'
 import { FileCriteria, FileSortByOption } from '../../../../files/domain/models/FileCriteria'
 import { DropdownButton, DropdownButtonItem } from 'dataverse-design-system'
 import { useTranslation } from 'react-i18next'
